결제, 장바구니, 관리자 수정

+++ Copy of ITLearnSite/src/cart/db/CartDAOImpl.java
@@ -168,7 +168,6 @@ public class CartDAOImpl implements CartDAO {
          closeConnection();
       }
    }
-
    // 장바구니 선택 삭제
    @Override
    public int delCart(int cart_num) {
@@ -186,21 +185,30 @@ public class CartDAOImpl implements CartDAO {
       return check;
    }
 
-   // 장바구니 전체 삭제
+   //장바구니 전체 삭제
    @Override
    public int delAllCart(String email) {
-      int check = 0;
-      try {
-         con = getConnection();
-         pstmt = con.prepareStatement("delete from cart_table where email = ?");
-         pstmt.setString(1, email);
-         check = pstmt.executeUpdate(); // 쿼리실행으로 삭제된 레코드 수 반환
-      } catch (Exception e) {
-         System.out.println("DelAllcart메소드에서 오류 :" + e);
-      } finally {
-         closeConnection();
-      }
-      return check;
+	   int dch = 0;
+	   try {
+		   con = getConnection();
+		   sql = "select pro_name from cart_table WHERE email = ?";
+		   pstmt = con.prepareStatement(sql);
+		   pstmt.setString(1, email);
+		   rs = pstmt.executeQuery();
+		   if(rs.next()){
+			   con = getConnection();
+	           pstmt = con.prepareStatement("delete from cart_table where email = ?");
+	           pstmt.setString(1, email);
+	           pstmt.executeUpdate();
+	           dch = 1;
+		   }  
+	        
+	      } catch (Exception e) {
+	         System.out.println("Delcart메소드에서 오류 :" + e);
+	      } finally {
+	         closeConnection();
+	      }
+	  return dch;
    }
    
    //장바구니 담긴 갯수 확인
